adding 6-sizes

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long_long;
 	float float_type;
 
-	printf("Size of char: %ld\n",sizeof(char_type));
-	printf("Size of int: %ld\n",sizeof(int_type)):
-	printf("Size of long: %ld\n",sizeof(long_int));
-	printf("Size of long long: %ld\n",sizeof(long_long));
-	printf("Size of float: %ld\n",sizeof(float_type));
+	printf("Size of char: %ld\n",(unsigned int)sizeof(char_type));
+	printf("Size of int: %ld\n",(unsigned int)sizeof(int_type)):
+	printf("Size of long: %ld\n",(unsigned int)sizeof(long_int));
+	printf("Size of long long: %ld\n",(unsigned int)sizeof(long_long));
+	printf("Size of float: %ld\n",(unsigned int)sizeof(float_type));
 	return (0);
 }
